Add server setup
Created a basic server file (`server.js`) with essential middleware including Express, Morgan for logging, Body Parser for parsing request bodies, CORS for handling cross-origin requests, and Mongoose for MongoDB integration. Also, defined a simple wildcard endpoint to acknowledge successful connection to the node.js API for the customer-meeting-tracker application. The server is configured to run on the specified port or default to 8000.dotenv is used for environment variables.