Add MongoDB connection in server.js
Integrated code to connect to MongoDB in the server file (`server.js`). The application will now establish a connection to the specified database upon startup. Connection status will be logged to the console. Note that the actual database URL is sourced from environment variables. Also, using nodemon to avoid from hard refresh for the server side.